CHORE: change all values of double to bigDecimal, add colombia module.

+++ src/main/java/ms/hispam/budget/dto/ComponentAmount.java
@@ -2,6 +2,8 @@ package ms.hispam.budget.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -9,5 +11,6 @@ import lombok.*;
 @AllArgsConstructor
 public class ComponentAmount {
     private String component;
-    private double amount;
+    private BigDecimal amount;
+
 }
